Adds group starting/stopping.
Group starting is defined by a list of "auto_load" names.
Group stopping is all running instances.

It will group start via a command and on "on_ready".
The "on_ready" only triggers on the start of RedBot,
not the module.

It will group stop via a command and on "on_shutdown".
The shutdown will only trigger via the RedBot command, and
will not trigger when the module is reloaded.