Improves button interpretation.
The action of pressing and holding are converted to
numbers separately. This allows the holding to be
for decimal seconds.
It also no longer defaults to a single button press
if button is right, but the two other arguments
are nonsense.
This only seems fair that if one types it out
incorrectly that the do not get to default to
a single button push/hold.